requirements.txt
asgiref==3.7.2
Django==5.0.3
gunicorn==21.2.0
packaging==24.0
psycopg2-binary==2.9.9
sqlparse==0.4.4
celery==5.2.7
python-decouple==3.6
redis==4.3.4
Pillow==11.0.0
django-rest-framework==3.15.2
django-cors-headers==4.6.0
